+++ R/cumulativeprobplot_ranks.R
@@ -5,7 +5,7 @@
 #' @import ggplot2
 #'
 #' @param data A NxJ matrix of data, where N is the number of observations and J the number of treatments. This feature is designed to display results from a standard NMA study.
-#' @param mcmc MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.
+#' @param mcmc MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_raceNMA} function.
 #' @param names A vector of intervention names (optional)
 #'
 #' @return A ggplot of a cumulative ranking plot.
@@ -16,6 +16,7 @@
 #' @export
 cumulativeprobplot_ranks <- function(data=NULL, mcmc=NULL, names=NULL){
   if(!is.null(data)){
+    if(is.matrix(data)){data <- as.data.frame(data)}
     if(all(names(data)[1:3]==c("chain","iteration","K"))){stop("It seems that you have supplied mcmc_raceNMA output instead of data. Please update your input arguments.")}
     J <- ncol(data)
     if(is.null(names)){
@@ -35,6 +36,7 @@ cumulativeprobplot_ranks <- function(data=NULL, mcmc=NULL, names=NULL){
       labs(x="Posterior Rank",y="Cumulative Probability",color=NULL)
   }
   if(!is.null(mcmc)){
+    if(is.matrix(mcmc)){mcmc <- as.data.frame(mcmc)}
     if(any(names(mcmc)[1:3]!=c("chain","iteration","K"))){stop("It seems that you have supplied data instead of mcmc_raceNMA output. Please update your input arguments.")}
     J <- (ncol(mcmc)-3)/3
     posterior_mu <- mcmc[,grep("mu",names(mcmc))]